Fix: Errors introduced by making chatbot a package
